+++ stan/math/mix/prob/laplace_latent_rng.hpp
@@ -32,7 +32,7 @@ template <typename LLFunc, typename LLArgs, typename ThetaVec,
 inline auto laplace_latent_tol_rng(
     LLFunc&& L_f, LLArgs&& ll_args, ThetaVec&& theta_0,
     CovarFun&& covariance_function, CovarArgs&& covar_args,
-    const double tolerance, const int64_t max_num_steps,
+    const double tolerance, const int max_num_steps,
     const int hessian_block_size, const int solver,
     const int max_steps_line_search, RNG& rng, std::ostream* msgs) {
   const laplace_options ops{hessian_block_size, solver, max_steps_line_search,
