+++ stan/math/mix/prob/laplace_marginal.hpp
@@ -31,9 +31,8 @@ template <bool propto = false, typename LFun, typename LArgs, typename CovarFun,
 inline auto laplace_marginal_tol(
     LFun&& L_f, LArgs&& l_args, const ThetaVec& theta_0,
     CovarFun&& covariance_function, CovarArgs&& covar_args, double tolerance,
-    int64_t max_num_steps, const int hessian_block_size, const int solver,
+    int max_num_steps, const int hessian_block_size, const int solver,
     const int max_steps_line_search, std::ostream* msgs) {
-  // TEST: provisional signature to agree with parser.
   laplace_options ops{hessian_block_size, solver, max_steps_line_search,
                       tolerance, max_num_steps};
   return laplace_marginal_density(
